Move the derivation json code to a new file
The single `derivation.cc` was getting too big!

Nothing else is changed, this is a pure move.
